requirements.txt
requests~2.19.1
base58~1.0.0
pynacl~1.2.1